+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#233](https://github.com/os2display/display-admin-client/pull/233)
+  - Fixed issue when clicking media archive element when media archive is not in multiple mode.
 - [#232](https://github.com/os2display/display-admin-client/pull/232)
   - Fixed time zone issue in playlist schedules.
 - [#231](https://github.com/os2display/display-admin-client/pull/231)